loco: migrate every crate from loco-rs 0.16 to 1.0.1
Fleet-wide major-version bump across all 348 back-end-with-loco crates,
applied mechanically via the new bin/loco-rs-1-migration tool:

- loco-rs 0.16 -> 1.0.1, sea-orm 1.1 -> 2.0, sea-orm-migration -> 2.0
- Feature rename auth_jwt/bg_pg -> auth/worker (the queue backend is now
  chosen at runtime via config/*.yaml, not a Cargo feature)
- Every id/*_id field (entities, controllers, hand-written modules) moved
  i32 -> i64, since loco 1.0's ColType::PkAuto now emits BIGINT
- Regenerated Cargo.lock for every crate, folding in the pending minor
  dependency bumps (tokio, uuid, insta, serial_test, sea-orm-migration
  patch) that were already sitting uncommitted in the tree

Also fixes three latent issues the version bump surfaced:

- Removed dead, unregistered Tera view_engine.rs initializers in
  architecture-decision-record, inpatient-clinical-note, and
  medical-operation-note, left over from initial scaffolding and
  contradicting these crates' own documented "no Tera" convention
- Worked around a cruet pluralization bug (to_plural("help") and
  to_plural("lpa_validity") return the word unchanged) that broke FK
  table-name resolution in agile-consulting-scorecard-for-hiring-help and
  united-kingdom-lasting-power-of-attorney-for-health-and-care-decisions
- Fixed 27 hand-written HTTP tests across 11 crates that requested
  /api/x/ with a trailing slash, which loco 1.0's routing no longer
  matches

Updated AGENTS/back-end-with-loco.md's version table and
bin/loco-config-refactor's canonical feature set (was still checking for
the pre-1.0 bg_pg/auth_jwt names), and regenerated docs/tools.md.

Verified: every one of the 348 crates passes cargo build --all-targets,
cargo test (against a live Postgres), and cargo clippy clean.

+++ forms/abdominal-aortic-aneurysm-screening/back-end-with-loco/src/abdominal_aortic_aneurysm_screening/models/_entities/abdominal_aortic_aneurysm_screening_grades.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub category: String,
     pub surveillance_band: String,
@@ -18,7 +18,7 @@ pub struct Model {
     #[sea_orm(column_type = "Double", nullable)]
     pub growth_cm: Option<f64>,
     pub graded_at: DateTimeWithTimeZone,
-    pub abdominal_aortic_aneurysm_screening_id: i32,
+    pub abdominal_aortic_aneurysm_screening_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
