+++ forms/abdominal-aortic-aneurysm-screening/back-end-with-loco/src/abdominal_aortic_aneurysm_screening/models/_entities/abdominal_aortic_aneurysm_screenings.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     #[sea_orm(column_type = "Text")]
     pub technician_name: String,
@@ -39,8 +39,8 @@ pub struct Model {
     pub incidental_findings: String,
     #[sea_orm(column_type = "Text")]
     pub result_note: String,
-    pub patient_id: i32,
-    pub clinician_id: Option<i32>,
+    pub patient_id: i64,
+    pub clinician_id: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
