+++ forms/advance-decision-to-refuse-treatment/back-end-with-loco/src/advance_decision_to_refuse_treatment/models/_entities/assessments.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub full_name: String,
@@ -121,7 +121,7 @@ pub struct Model {
     pub declaration_statement_agreed: String,
     #[sea_orm(column_type = "Text")]
     pub legal_notes: String,
-    pub patient_id: i32,
+    pub patient_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
