+++ forms/advance-statement-about-care/back-end-with-loco/src/advance_statement_about_care/models/_entities/assessments.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub full_name: String,
@@ -130,7 +130,7 @@ pub struct Model {
     pub review_date: Option<Date>,
     #[sea_orm(column_type = "Text")]
     pub review_notes: String,
-    pub patient_id: i32,
+    pub patient_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
