+++ forms/agile-checklist/back-end-with-loco/src/agile_checklist/models/_entities/agile_checklist_grade_flags.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub flag_id: String,
     pub category: String,
@@ -18,7 +18,7 @@ pub struct Model {
     pub triggering_items: String,
     pub description: String,
     pub suggested_action: String,
-    pub agile_checklist_grade_id: i32,
+    pub agile_checklist_grade_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
