+++ forms/agile-consulting-scorecard-for-hiring-help/back-end-with-loco/src/agile_consulting_scorecard_for_hiring_help/controllers/agile_consulting_scorecard_for_hiring_help.rs
@@ -9,8 +9,8 @@ use crate::models::_entities::agile_consulting_scorecard_for_hiring_helps::{Acti
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Params {
     pub deleted_at: Option<DateTimeWithTimeZone>,
-    pub organization_id: i32,
-    pub respondent_id: i32,
+    pub organization_id: i64,
+    pub respondent_id: i64,
     pub status: String,
     pub assessment_date: Date,
     pub submitted_at: Option<DateTimeWithTimeZone>,
@@ -93,7 +93,7 @@ impl Params {
       }
 }
 
-async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
+async fn load_item(ctx: &AppContext, id: i64) -> Result<Model> {
     let item = Entity::find_by_id(id).one(&ctx.db).await?;
     item.ok_or_else(|| Error::NotFound)
 }
@@ -115,7 +115,7 @@ pub async fn add(State(ctx): State<AppContext>, Json(params): Json<Params>) -> R
 
 #[debug_handler]
 pub async fn update(
-    Path(id): Path<i32>,
+    Path(id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
 ) -> Result<Response> {
@@ -127,13 +127,13 @@ pub async fn update(
 }
 
 #[debug_handler]
-pub async fn remove(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn remove(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     load_item(&ctx, id).await?.delete(&ctx.db).await?;
     format::empty()
 }
 
 #[debug_handler]
-pub async fn get_one(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn get_one(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     format::json(load_item(&ctx, id).await?)
 }
