+++ forms/agile-consulting-scorecard-for-hiring-help/back-end-with-loco/src/agile_consulting_scorecard_for_hiring_help/models/_entities/agile_consulting_scorecard_for_hiring_helps.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub assessment_date: Date,
@@ -63,8 +63,8 @@ pub struct Model {
     pub p12_done: Option<bool>,
     #[sea_orm(column_type = "Text")]
     pub p12_evidence: String,
-    pub organization_id: i32,
-    pub respondent_id: i32,
+    pub organization_id: i64,
+    pub respondent_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
