+++ forms/agile-principles-assessment/back-end-with-loco/src/agile_principles_assessment/models/_entities/agile_principles_assessment_grade_flags.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub flag_id: String,
     pub category: String,
@@ -17,7 +17,7 @@ pub struct Model {
     pub principle_number: Option<i32>,
     pub description: String,
     pub suggested_action: String,
-    pub agile_principles_assessment_grade_id: i32,
+    pub agile_principles_assessment_grade_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
