+++ forms/agile-principles-assessment/back-end-with-loco/src/agile_principles_assessment/models/_entities/agile_principles_assessment_grade_rules.rs
@@ -9,14 +9,14 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub rule_id: String,
     pub principle_number: i32,
     pub band: String,
     pub principle_slug: String,
     pub description: String,
-    pub agile_principles_assessment_grade_id: i32,
+    pub agile_principles_assessment_grade_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
