+++ forms/alcohol-use-disorders-identification-test-consumption/back-end-with-loco/src/alcohol_use_disorders_identification_test_consumption/models/_entities/audit_c_grades.rs
@@ -9,13 +9,13 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub total_score: Option<i32>,
     pub risk_band: String,
     pub positive_screen: String,
     pub graded_at: DateTimeWithTimeZone,
-    pub audit_c_id: i32,
+    pub audit_c_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
