+++ forms/allergy-assessment/back-end-with-loco/src/allergy_assessment/models/_entities/assessment_drug_allergies.rs
@@ -9,12 +9,12 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub has_drug_allergies: String,
     #[sea_orm(column_type = "Text")]
     pub cross_reactivity_concerns: String,
-    pub assessment_id: i32,
+    pub assessment_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
