+++ forms/allergy-assessment/back-end-with-loco/src/allergy_assessment/models/_entities/assessments.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub age_of_onset: Option<i32>,
@@ -54,7 +54,7 @@ pub struct Model {
     pub training_details: String,
     #[sea_orm(column_type = "Text")]
     pub follow_up_schedule: String,
-    pub patient_id: i32,
+    pub patient_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
