+++ forms/allergy-skin-test-request/back-end-with-loco/src/allergy_skin_test_request/controllers/allergy_skin_test_request_grade.rs
@@ -9,7 +9,7 @@ use crate::models::_entities::allergy_skin_test_request_grades::{ActiveModel, En
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Params {
     pub deleted_at: Option<DateTimeWithTimeZone>,
-    pub allergy_skin_test_request_id: i32,
+    pub allergy_skin_test_request_id: i64,
     pub appropriateness_score: Option<i32>,
     pub appropriateness_band: String,
     pub validity_safety_band: String,
@@ -39,7 +39,7 @@ impl Params {
       }
 }
 
-async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
+async fn load_item(ctx: &AppContext, id: i64) -> Result<Model> {
     let item = Entity::find_by_id(id).one(&ctx.db).await?;
     item.ok_or_else(|| Error::NotFound)
 }
@@ -61,7 +61,7 @@ pub async fn add(State(ctx): State<AppContext>, Json(params): Json<Params>) -> R
 
 #[debug_handler]
 pub async fn update(
-    Path(id): Path<i32>,
+    Path(id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
 ) -> Result<Response> {
@@ -73,13 +73,13 @@ pub async fn update(
 }
 
 #[debug_handler]
-pub async fn remove(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn remove(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     load_item(&ctx, id).await?.delete(&ctx.db).await?;
     format::empty()
 }
 
 #[debug_handler]
-pub async fn get_one(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn get_one(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     format::json(load_item(&ctx, id).await?)
 }
