+++ forms/ambulatory-blood-pressure-test-result/back-end-with-loco/src/ambulatory_blood_pressure_test_result/controllers/ambulatory_blood_pressure_test_result.rs
@@ -9,8 +9,8 @@ use crate::models::_entities::ambulatory_blood_pressure_test_results::{ActiveMod
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Params {
     pub deleted_at: Option<DateTimeWithTimeZone>,
-    pub patient_id: i32,
-    pub clinician_id: i32,
+    pub patient_id: i64,
+    pub clinician_id: i64,
     pub originating_request_reference: String,
     pub monitoring_type: String,
     pub report_status: String,
@@ -79,7 +79,7 @@ impl Params {
       }
 }
 
-async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
+async fn load_item(ctx: &AppContext, id: i64) -> Result<Model> {
     let item = Entity::find_by_id(id).one(&ctx.db).await?;
     item.ok_or_else(|| Error::NotFound)
 }
@@ -101,7 +101,7 @@ pub async fn add(State(ctx): State<AppContext>, Json(params): Json<Params>) -> R
 
 #[debug_handler]
 pub async fn update(
-    Path(id): Path<i32>,
+    Path(id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
 ) -> Result<Response> {
@@ -113,13 +113,13 @@ pub async fn update(
 }
 
 #[debug_handler]
-pub async fn remove(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn remove(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     load_item(&ctx, id).await?.delete(&ctx.db).await?;
     format::empty()
 }
 
 #[debug_handler]
-pub async fn get_one(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn get_one(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     format::json(load_item(&ctx, id).await?)
 }
