+++ forms/anaesthetic-record/back-end-with-loco/src/anaesthetic_record/models/_entities/anaesthetic_record_drug_administrations.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub drug_name: String,
     #[sea_orm(column_type = "Double", nullable)]
@@ -18,7 +18,7 @@ pub struct Model {
     pub route: String,
     pub category: String,
     pub administered_at: Option<DateTimeWithTimeZone>,
-    pub anaesthetic_record_id: i32,
+    pub anaesthetic_record_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
