+++ forms/anaesthetic-record/back-end-with-loco/src/anaesthetic_record/models/_entities/anaesthetic_record_grade_rules.rs
@@ -9,14 +9,14 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub rule_id: String,
     pub category: String,
     pub criticality: String,
     pub satisfied: bool,
     pub description: String,
-    pub anaesthetic_record_grade_id: i32,
+    pub anaesthetic_record_grade_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
