+++ forms/anaesthetic-record/back-end-with-loco/src/anaesthetic_record/models/_entities/anaesthetic_record_grades.rs
@@ -9,12 +9,12 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub completeness_percent: i32,
     pub graded_at: DateTimeWithTimeZone,
-    pub anaesthetic_record_id: i32,
+    pub anaesthetic_record_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
