+++ forms/anaesthetic-record/back-end-with-loco/src/anaesthetic_record/models/_entities/anaesthetic_record_intra_operative_events.rs
@@ -9,13 +9,13 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub event_type: String,
     pub occurred_at: Option<DateTimeWithTimeZone>,
     #[sea_orm(column_type = "Text")]
     pub management: String,
-    pub anaesthetic_record_id: i32,
+    pub anaesthetic_record_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
