+++ forms/anion-gap-calculator/back-end-with-loco/src/anion_gap_calculator/models/_entities/anion_gap_calculator_grades.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     #[sea_orm(column_type = "Double", nullable)]
     pub anion_gap: Option<f64>,
@@ -17,7 +17,7 @@ pub struct Model {
     pub corrected_anion_gap: Option<f64>,
     pub classification: String,
     pub graded_at: DateTimeWithTimeZone,
-    pub anion_gap_calculator_id: i32,
+    pub anion_gap_calculator_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
