+++ forms/apgar-score/back-end-with-loco/src/apgar_score/models/_entities/apgar_score_grade_rules.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub rule_id: String,
     pub timepoint_minutes: Option<i32>,
@@ -17,7 +17,7 @@ pub struct Model {
     pub points: Option<i32>,
     pub category: String,
     pub description: String,
-    pub apgar_score_grade_id: i32,
+    pub apgar_score_grade_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
