+++ forms/apgar-score/back-end-with-loco/src/apgar_score/models/_entities/apgar_score_grades.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub total_one_minute: Option<i32>,
     pub total_five_minute: Option<i32>,
@@ -20,7 +20,7 @@ pub struct Model {
     pub summary_band: String,
     pub trend: String,
     pub graded_at: DateTimeWithTimeZone,
-    pub apgar_score_id: i32,
+    pub apgar_score_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
