+++ forms/apgar-score/back-end-with-loco/src/apgar_score/models/_entities/apgar_score_timepoints.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub timepoint_minutes: Option<i32>,
     pub appearance: String,
@@ -19,7 +19,7 @@ pub struct Model {
     pub respiration: String,
     pub total: Option<i32>,
     pub band: String,
-    pub apgar_score_id: i32,
+    pub apgar_score_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
