+++ forms/arc42/back-end-with-loco/src/arc42/models/_entities/stakeholders.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub ordinal: i32,
     #[sea_orm(column_type = "Text")]
@@ -18,7 +18,7 @@ pub struct Model {
     pub role: String,
     #[sea_orm(column_type = "Text")]
     pub concerns: String,
-    pub arc42_documentation_id: i32,
+    pub arc42_documentation_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
